refactor: extract tab display/git-root cluster from window-tabs (#281)
Sweep finding B2, step 1 of the incremental god-store decomposition:
the ~130-line labeling cluster (git-root cache + async gitRepoRoot IPC,
VS Code-style single-pane disambiguation, multi-pane title joining)
existed purely to decorate tab titles yet lived inside
createWindowTabsManager. It now lives in state/tab-display.svelte.ts,
reading tabs/paths through injected getters so every derivation
recomputes exactly as before. extractFolderName moves to its true home
in domain/tab-title.ts (re-exported from window-tabs for importers).

Behavior pinned by tests/state/window-tabs.test.ts (63) and the
tab-title/pane-tabs E2E suites — all green.
